# docs/swagger.yaml
basePath: /
...
host: localhost:3000
info:
  contact:
    email: support@aygit.com
    name: API Destek
    url: http://www.aygit.com/destek
  description: Bu servis, NES Özel Entegratör üzerinden e-fatura, e-arşiv ve ürün
    yönetim süreçlerini yönetir.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Aygıt Muhasebe Entegrasyonu API
  version: "1.0"
paths:
...
  /api/v1/{type}/notifications/dynamicrules:
    get:
      consumes:
      - application/json
      description: Kuralları listeler
      parameters:
      - description: 'Voucher Type: emm or esmm'
        in: path
        name: type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Get Dynamic Rules
      tags:
      - Notifications
    post:
      consumes:
      - application/json
      description: Kural oluşturur
      parameters:
      - description: 'Voucher Type: emm or esmm'
        in: path
        name: type
        required: true
        type: string
      - description: Rule Data
        in: body
        name: payload
        required: true
        schema: {}
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Create Dynamic Rule
      tags:
      - Notifications
  /api/v1/{type}/notifications/dynamicrules/{id}:
    delete:
      consumes:
      - application/json
      description: Kural siler
      parameters:
      - description: 'Voucher Type: emm or esmm'
        in: path
        name: type
        required: true
        type: string
      - description: Rule ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Delete Dynamic Rule
      tags:
      - Notifications
    get:
      consumes:
      - application/json
      description: Sorgulanan kuralı getirir
      parameters:
      - description: 'Voucher Type: emm or esmm'
        in: path
        name: type
        required: true
        type: string
      - description: Rule ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Get Dynamic Rule by ID
      tags:
      - Notifications
    put:
      consumes:
      - application/json
      description: Kural günceller
      parameters:
      - description: 'Voucher Type: emm or esmm'
        in: path
        name: type
        required: true
        type: string
      - description: Rule ID
        in: path
        name: id
        required: true
        type: string
      - description: Rule Data
        in: body
        name: payload
        required: true
        schema: {}
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Update Dynamic Rule
      tags:
      - Notifications
  /api/v1/{type}/outgoing/reportmodule/columns:
    get:
      consumes:
      - application/json
      description: Kolonları listeler
      parameters:
      - description: 'Voucher Type: emm or esmm'
        in: path
        name: type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Get Columns
      tags:
      - Outgoing
  /api/v1/{type}/outgoing/reportmodule/reports:
    get:
      consumes:
      - application/json
      description: Rapor listeler
      parameters:
      - description: 'Voucher Type: emm or esmm'
        in: path
        name: type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Get Reports
      tags:
      - Outgoing
    post:
      consumes:
      - application/json
      description: Rapor oluşturur
      parameters:
      - description: 'Voucher Type: emm or esmm'
        in: path
        name: type
        required: true
        type: string
      - description: Data
        in: body
        name: payload
        required: true
        schema: {}
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Create Report
      tags:
      - Outgoing
  /api/v1/{type}/outgoing/reportmodule/reports/{id}/download:
    get:
      consumes:
      - application/json
      description: Rapor indirir
      parameters:
      - description: 'Voucher Type: emm or esmm'
        in: path
        name: type
        required: true
        type: string
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Download Report
      tags:
      - Outgoing
  /api/v1/{type}/outgoing/reportmodule/templates:
    get:
      consumes:
      - application/json
      description: Şablonları listeler
      parameters:
      - description: 'Voucher Type: emm or esmm'
        in: path
        name: type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Get Report Templates
      tags:
      - Outgoing
    post:
      consumes:
      - application/json
      description: Rapor şablonu oluşturur
      parameters:
      - description: 'Voucher Type: emm or esmm'
        in: path
        name: type
        required: true
        type: string
      - description: Data
        in: body
        name: payload
        required: true
        schema: {}
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Create Report Template
      tags:
      - Outgoing
  /api/v1/{type}/outgoing/reportmodule/templates/{id}:
    delete:
      consumes:
      - application/json
      description: Rapor Şablonunu siler
      parameters:
      - description: 'Voucher Type: emm or esmm'
        in: path
        name: type
        required: true
        type: string
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Delete Report Template
      tags:
      - Outgoing
    get:
      consumes:
      - application/json
      description: Sorgulanan şablonu getirir
      parameters:
      - description: 'Voucher Type: emm or esmm'
        in: path
        name: type
        required: true
        type: string
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Get Report Template By ID
      tags:
      - Outgoing
    put:
      consumes:
      - application/json
      description: Rapor şablonunu günceller
      parameters:
      - description: 'Voucher Type: emm or esmm'
        in: path
        name: type
        required: true
        type: string
      - description: ID
        in: path
        name: id
        required: true
        type: string
      - description: Data
        in: body
        name: payload
        required: true
        schema: {}
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Update Report Template
      tags:
      - Outgoing
...
  /v1/envelopes/{instanceIdentifier}/query:
    get:
      parameters:
      - description: Instance Identifier
        in: path
        name: instanceIdentifier
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Zarf Durum Sorgular
      tags:
      - Envelopes
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"